feat: Implement MarketViewModel
The view model for the market feature handle two main scenarios.
- Handle incoming actions (intents)
- Update Ui state

The first action implemented is to load the asset list.
Once is retrieved, the Flow that keeps track of the UI state is
updated.
At a later stage the UI will observe the UI flow to update the
interface accordingly.